// src/app/models/development/statusPlannedTask.ts
export enum StatusPlannedTask {
    NOT_READY = 'NOT_READY',
    NOT_STARTED = 'NOT_STARTED',
    PUBLISHED = 'PUBLISHED',
    REQUESTS_FINISHED = 'REQUESTS_FINISHED',
    PERFORMER_SELECTED = 'PERFORMER_SELECTED',
    IN_PROGRESS = 'IN_PROGRESS',
    FINISHED = 'FINISHED',
  
    ERROR = 'ERROR'
}